backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.10.0
pydantic-settings
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ML/AI - Using latest versions with pre-built wheels for Windows
sentence-transformers==2.7.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Utilities
python-dotenv==1.0.0
